api/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
openpyxl
pydantic
torch
transformers
requests
python-dotenv
python-multipart